# HttpApi: Remove 'requests' to use 'urllib3' directly