--- a/all-requirements.txt
+++ b/all-requirements.txt
@@ -10,3 +10,4 @@ plyvel>=0.9
 PyYAML>=3.10
 PyECLib>=1.2.0
 futures>=3.0.5
+urllib3>=1.13.1