To improve the event agent, I change `HttpApi` class. Because `HttpApi` use `requests` only for connections pools. To manage connections pools, `requests` use `urllib3` with `PoolManager`. Now, I call `PoolManager` directly.

With the benchmark `oio-event-benchmark`, we increase the speed by more than 60%.
- before:
```shell
$ ./build/tools/event-benchmark/oio-event-benchmark OPENIO CHUNK_NEW -q
Linking fake rawx address with the fake service address...
Starting the fake service...
Sending 100 fake events
100 events sent (errors: 0) in 0.105229 seconds, 950.308375 events/sec
Sending 950 fake events
950 events sent (errors: 0) in 0.641872 seconds, 1480.045866 events/sec
Sending 1480 fake events
1480 events sent (errors: 0) in 1.021468 seconds, 1448.895120 events/sec
Sending 1448 fake events
1448 events sent (errors: 0) in 1.007678 seconds, 1436.966968 events/sec
Sending 1436 fake events
1436 events sent (errors: 0) in 1.010318 seconds, 1421.334669 events/sec
Sending 1421 fake events
1421 events sent (errors: 0) in 0.981576 seconds, 1447.671907 events/sec
Sending 1447 fake events
1447 events sent (errors: 0) in 1.039129 seconds, 1392.512383 events/sec
Sending 1392 fake events
1392 events sent (errors: 0) in 0.964230 seconds, 1443.638966 events/sec
Sending 1443 fake events
^CStopping the fake service...
1443 events sent (errors: 0) in 0.991372 seconds, 1455.558559 events/sec
Cleaning...
```
- after:
```shell
$ ./build/tools/event-benchmark/oio-event-benchmark OPENIO CHUNK_NEW -q
Linking fake rawx address with the fake service address...
Starting the fake service...
Sending 100 fake events
100 events sent (errors: 0) in 0.064317 seconds, 1554.798887 events/sec
Sending 1554 fake events
1554 events sent (errors: 0) in 0.651858 seconds, 2383.954788 events/sec
Sending 2383 fake events
2383 events sent (errors: 0) in 1.044297 seconds, 2281.917884 events/sec
Sending 2281 fake events
2281 events sent (errors: 0) in 0.946070 seconds, 2411.026668 events/sec
Sending 2411 fake events
2411 events sent (errors: 0) in 1.036022 seconds, 2327.170659 events/sec
Sending 2327 fake events
2327 events sent (errors: 0) in 0.967479 seconds, 2405.220165 events/sec
Sending 2405 fake events
2405 events sent (errors: 0) in 0.995019 seconds, 2417.039273 events/sec
Sending 2417 fake events
2417 events sent (errors: 0) in 1.018104 seconds, 2374.020729 events/sec
Sending 2374 fake events
2374 events sent (errors: 0) in 1.033525 seconds, 2296.993300 events/sec
Sending 2296 fake events
2296 events sent (errors: 0) in 0.955092 seconds, 2403.956896 events/sec
Sending 2403 fake events
^CStopping the fake service...
2403 events sent (errors: 0) in 1.044140 seconds, 2301.415519 events/sec
Cleaning...
```